refactor: avoid cyclic dependencies
