+++ src/Collapse/Collapse.ts
