+++ src/Collapse/index.tsx
@@ -1,2 +1,2 @@
-export * from './common';
+export * from './Collapse';
 export * from './SingleCollapse';
