+++ src/Modal/Modal.tsx
@@ -1,0 +1,4 @@
+import { Modal as AntdModal, ModalProps as AntdModalProps } from 'antd';
+
+export const Modal: typeof AntdModal = AntdModal;
+export type ModalProps = AntdModalProps;
